git-svn-id: http://xe-core.googlecode.com/svn/trunk@734 201d5d3c-b55e-5fd7-737f-ddc643e51545

+++ classes/db/DBMysql.class.php
@@ -24,7 +24,7 @@
          **/
         var $column_type = array(
             'bignumber' => 'bigint',
-            'number' => 'int',
+            'number' => 'bigint',
             'varchar' => 'varchar',
             'char' => 'char',
             'text' => 'text',
